refactor activations
