+++ modulus/sym/models/multiplicative_filter_net.py
@@ -18,9 +18,8 @@ from typing import Optional, List, Dict, Tuple, Union
 import torch.nn as nn
 from torch import Tensor
 
-import modulus.sym.models.layers as layers
 from modulus.models.layers import FCLayer, FourierFilter, GaborFilter
-from modulus.sym.models.layers import get_activation_fn
+from modulus.sym.models.activation import Activation, get_activation_fn
 from modulus.sym.models.arch import Arch
 from modulus.sym.key import Key
 from modulus.sym.constants import NO_OP_NORM
@@ -58,7 +57,7 @@ class MultiplicativeFilterNetArch(Arch):
         Number of hidden layers of the model.
     skip_connections : bool = False
         If true then apply skip connections every 2 hidden layers.
-    activation_fn : layers.Activation = layers.Activation.SILU
+    activation_fn : Activation = Activation.SILU
         Activation function used by network.
     filter_type : FilterType = FilterType.FOURIER
         Filter type for multiplicative filter network, (Fourier or Gabor).
@@ -82,7 +81,7 @@ class MultiplicativeFilterNetArch(Arch):
         layer_size: int = 512,
         nr_layers: int = 6,
         skip_connections: bool = False,
-        activation_fn=layers.Activation.IDENTITY,
+        activation_fn=Activation.IDENTITY,
         filter_type: Union[FilterType, str] = FilterType.FOURIER,
         weight_norm: bool = True,
         input_scale: float = 10.0,
