+++ modulus/sym/models/multiscale_fourier_net.py
@@ -18,9 +18,8 @@ import torch
 import torch.nn as nn
 from torch import Tensor
 
-import modulus.sym.models.layers as layers
 from modulus.models.layers import FCLayer, FourierLayer
-from modulus.sym.models.layers import get_activation_fn
+from modulus.sym.models.activation import Activation, get_activation_fn
 from modulus.sym.models.arch import Arch
 from modulus.sym.key import Key
 
@@ -53,7 +52,7 @@ class MultiscaleFourierNetArch(Arch):
     frequencies_params : Tuple[Tuple[str, List[float]],...] = (("axis", [i for i in range(10)]),)
         Same as `frequencies` except these are used for encodings
         on any inputs not in the list `['x', 'y', 'z', 't']`.
-    activation_fn : layers.Activation = layers.Activation.SILU
+    activation_fn : Activation = Activation.SILU
         Activation function used by network.
     layer_size : int = 512
         Layer size for every hidden layer of the model.
@@ -75,7 +74,7 @@ class MultiscaleFourierNetArch(Arch):
         detach_keys: List[Key] = [],
         frequencies=(("axis", [i for i in range(10)]),),
         frequencies_params=(("axis", [i for i in range(10)]),),
-        activation_fn=layers.Activation.SILU,
+        activation_fn=Activation.SILU,
         layer_size: int = 512,
         nr_layers: int = 6,
         skip_connections: bool = False,
